Fix plaintext header definition

+++ src/insets/InsetArgument.h
@@ -49,7 +49,7 @@ public:
 	///
 	void latex(otexstream &, OutputParams const &) const { }
 	///
-	int plaintext(odocstringstream &, OutputParams const &, int) const { return 0; }
+	int plaintext(odocstringstream &, OutputParams const &, size_t) const { return 0; }
 	///
 	int docbook(odocstream &, OutputParams const &) const { return 0; }
 	///
